move to hybrid_in.cpp
